requirements.txt
tensorflow==2.12.0
keras==2.12.0
Pillow==9.5.0
numpy==1.23.5
scikit-learn==1.2.2
opencv-python==4.7.0.72
tk
